+++ app/src/main/java/hannah/bd/getitwrite/views/account/StatsView.kt
@@ -48,9 +48,9 @@ fun StatsSection(user: User) {
             )
         }
         user.frequencey?.let { timeInSeconds ->
-            val days = TimeUnit.SECONDS.toDays(timeInSeconds)
-            val hours = TimeUnit.SECONDS.toHours(timeInSeconds) % 24
-            val minutes = TimeUnit.SECONDS.toMinutes(timeInSeconds) % 60
+            val days = TimeUnit.SECONDS.toDays(timeInSeconds.toLong())
+            val hours = TimeUnit.SECONDS.toHours(timeInSeconds.toLong()) % 24
+            val minutes = TimeUnit.SECONDS.toMinutes(timeInSeconds.toLong()) % 60
             val seconds = timeInSeconds % 60
 
             val humanReadableTime = buildString {
@@ -97,9 +97,9 @@ fun StatsView(user: User, navController: NavHostController) {
                 Text(text = "Not enough data to create a graph.")
             }
             user.frequencey?.let { timeInSeconds ->
-                val days = TimeUnit.SECONDS.toDays(timeInSeconds)
-                val hours = TimeUnit.SECONDS.toHours(timeInSeconds) % 24
-                val minutes = TimeUnit.SECONDS.toMinutes(timeInSeconds) % 60
+                val days = TimeUnit.SECONDS.toDays(timeInSeconds.toLong())
+                val hours = TimeUnit.SECONDS.toHours(timeInSeconds.toLong()) % 24
+                val minutes = TimeUnit.SECONDS.toMinutes(timeInSeconds.toLong()) % 60
                 val seconds = timeInSeconds % 60
 
                 val humanReadableTime = buildString {
